requirements.txt
firebase_admin
flask
loguru
Flask-WTF
Flask-Bootstrap4